refactor refresh sharding metadata. move dialect metadata from sharding-jdbc-core to sharding-core, to ensure proxy-sharding can simply invoke.
