+++ sharding-core/src/main/java/io/shardingjdbc/core/metadata/dialect/MySQLShardingMetaDataHandler.java
@@ -15,10 +15,9 @@
  * </p>
  */
 
-package io.shardingjdbc.core.jdbc.metadata.dialect;
+package io.shardingjdbc.core.metadata.dialect;
 
 import io.shardingjdbc.core.metadata.ColumnMetaData;
-
 import javax.sql.DataSource;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -32,22 +31,23 @@ import java.util.List;
  * @author panjuan
  * @author zhaojun
  */
-public final class H2ShardingMetaDataHandler extends ShardingMetaDataHandler {
+public final class MySQLShardingMetaDataHandler extends ShardingMetaDataHandler {
     
-    public H2ShardingMetaDataHandler(final DataSource dataSource, final String actualTableName) {
+    public MySQLShardingMetaDataHandler(final DataSource dataSource, final String actualTableName) {
         super(dataSource, actualTableName);
     }
-    
+
     @Override
     public List<ColumnMetaData> geColumnMetaInternal(final Statement statement) throws SQLException {
         List<ColumnMetaData> result = new LinkedList<>();
-        ResultSet resultSet = statement.getConnection().getMetaData().getTables(null, null, getActualTableName(), null);
+        statement.executeQuery(String.format("show tables like '%s'", getActualTableName()));
+        ResultSet resultSet = statement.getResultSet();
 
         if (resultSet.next()) {
-            statement.executeQuery(String.format("show columns from %s;", getActualTableName()));
+            statement.executeQuery(String.format("desc %s;", getActualTableName()));
             resultSet = statement.getResultSet();
             while (resultSet.next()) {
-                result.add(new ColumnMetaData(resultSet.getString("FIELD"), resultSet.getString("TYPE"), resultSet.getString("KEY")));
+                result.add(new ColumnMetaData(resultSet.getString("Field"), resultSet.getString("Type"), resultSet.getString("Key")));
             }
         }
         return result;
